+++ spec/lib/karafka/instrumentation/monitor_spec.rb
@@ -49,7 +49,7 @@ RSpec.describe Karafka::Instrumentation::Monitor do
 
   describe '#available_events' do
     it 'expect to include registered events' do
-      expect(monitor.available_events.size).to eq 17
+      expect(monitor.available_events.size).to eq 18
     end
 
     it { expect(monitor.available_events).to include 'connection.listener.fetch_loop.error' }