+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - #340 - Procline instrumentation for a nicer process name
 - #342 - Change default for `fetcher_max_queue_size` from `100` to `10` to lower max memory usage
 - #345 - Cleanup exceptions names
+- #341 - Split conneciton delegator into batch delegator and single_delegator
 
 ## 1.2.4
 - #332 - Fetcher for max queue size